% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymDist.R
\name{asymDist1D}
\alias{asymDist1D}
\title{Calculate the asymmetric distance for a dataframe of aligned sequences based on one reference sequence}
\usage{
asymDist1D(
  Alignment,
  DeltaMatrix = capasydis::DeltaMatrix,
  PosRef = 1,
  Ncores = 1
)
}
\arguments{
\item{Alignment}{the alignment provided as a data frame of sequences (rows) by aligned positions (columns).}

\item{DeltaMatrix}{the substitution matrix used to define the distance between pairs of bases being compared.}

\item{PosRef}{the index (row number) of the sequence to be used as the reference.}

\item{Ncores}{the number of cores chosen for parallelization}
}
\value{
A list of class "list", and "asymdist1D" containing the \strong{minScore}, \strong{maxScore}, range of the scores \strong{RangeScore}, and the asymmetric distances (\strong{asymdist}) for each sequence using the PosRef sequence as reference sequence.
}
\description{
Calculate the asymmetric distance for a dataframe of aligned sequences based on one reference sequence
}
\examples{
Alignment_df <-
 data.frame(rbind(
   Reference=c("A","C","G","T","A","T","G","C","A","T"),
   S1=c("A","A","G","T","A","T","G","C","A","T"),
   S2=c("A","C","G","T","C","T","G","C","A","T"),
   S3=c("A","C","G","T","A","T","G","C","T","T"))
 )
asymDist1D(Alignment=Alignment_df,DeltaMatrix=DeltaMatrix,PosRef=1)
}
