% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymDist2D.R
\name{asymDist2D}
\alias{asymDist2D}
\title{Calculate the asymmetric distance for a dataframe of aligned sequences based on two reference sequences}
\usage{
asymDist2D(
  Alignment,
  DeltaMatrix = capasydis::DeltaMatrix,
  PosRef1 = 1,
  PosRef2 = 4,
  Ncores = 2
)
}
\arguments{
\item{Alignment}{the alignment provided as a data frame of sequences (rows) by aligned positions (columns).}

\item{DeltaMatrix}{the substitution matrix used to define the distance between pairs of bases being compared.}

\item{PosRef1}{the index (row number) of the first sequence to be used as the reference.}

\item{PosRef2}{the index (row number) of the second sequence to be used as the reference.}

\item{Ncores}{the number of cores chosen for parallelization}
}
\value{
A list of class "asymdist2D" containing the names of sequences, the distance of each sequence to sequence given in position 1 (disttoRefPos1), the distance of each sequence to sequence given in position 2 (disttoRefPos2).
}
\description{
Calculate the asymmetric distance for a dataframe of aligned sequences based on two reference sequences
}
\examples{
Alignment_df <-
 data.frame(rbind(
   Reference=c("A","C","G","T","A","T","G","C","A","T"),
   S1=c("A","A","G","T","A","T","G","C","A","T"),
   S2=c("A","C","G","T","C","T","G","C","A","T"),
   S3=c("A","C","G","T","A","T","G","C","T","T"))
 )

asymDist2D(Alignment=Alignment_df,DeltaMatrix=capasydis::DeltaMatrix,PosRef1=1,PosRef2=4,Ncores=2)
}
