% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcDelta2seqs.R
\name{CalcDelta2seqs}
\alias{CalcDelta2seqs}
\title{Calculate the delta vector between two vectors of characters}
\usage{
CalcDelta2seqs(REF, SEQ, DeltaVector)
}
\arguments{
\item{REF}{the reference vector of characters (e.g. bases)}

\item{SEQ}{the query vector of characters (e.g. bases)}

\item{DeltaVector}{the name vector of delta value: e.g. DeltaMatrix}
}
\value{
for each pair in REF and SEQ, the delta value is calculated.
}
\description{
Calculate the delta vector between two vectors of characters
}
\examples{
CalcDelta2seqs(REF=c("A","A"),SEQ=c("A","A"),DeltaVector = capasydis::DeltaMatrix) # 0 0
CalcDelta2seqs(REF=c("A","A"),SEQ=c("G","C"),DeltaVector = capasydis::DeltaMatrix) # 0.01 0.06


}
