% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckMSA.R
\name{CheckMSA}
\alias{CheckMSA}
\title{Check for absence of unknown or of wobble bases in all the sequences of an alignment}
\usage{
CheckMSA(Alignment, type = "N")
}
\arguments{
\item{Alignment}{the sequences as a data frame of aligned sequences (rows) of single chars}

\item{type}{the type of bases to check for "N" for unknown only, "W" for wobble bases (not "A","T","G","C"), or "b" for both}
}
\value{
A list with two slots indicating the sequences (rows) where unknowns or wobble bases were found.
}
\description{
Check for absence of unknown or of wobble bases in all the sequences of an alignment
}
\examples{
AL <- rbind(
S1=c("A","T","G","C","A","T","G","C"),
S2=c("A","N","G","C","A","T","G","C"),
S3=c("A","T","G","C","A","T","G","C"),
S4=c("A","T","G","W","A","T","G","C"),
S5=c("A","T","G","W","A","T","G","N")
)
C <- CheckMSA(Alignment = AL)
C$N # 2 5
C$W # NULL
C <- CheckMSA(Alignment = AL,type="W")
C$N # NULL
C$W # 2 4 5
C <- CheckMSA(Alignment = AL,type="b")
C$N # 2 5
C$W # 2 4 5
C <- CheckMSA(Alignment = AL[c(1,3),],type="b")
C$N # 0
C$W # 0

}
