% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostDivergentScore.R
\name{MostDivergentScore}
\alias{MostDivergentScore}
\title{Calculate the most divergent score and sequence for one sequence of characters}
\usage{
MostDivergentScore(
  SEQ,
  MatrixMaxDivergence = MatrixMaxDivergence,
  DeltaMatrix = DeltaMatrix
)
}
\arguments{
\item{SEQ}{the sequence of characters of interest (e.g. SEQ=REF1)}

\item{MatrixMaxDivergence}{the vector of maximum divergence}

\item{DeltaMatrix}{the delta matrix}
}
\value{
A list is produced with  \code{max.delta.vec}, as the maximum diverging delta for each character in the provided sequence; \code{Score} the associated score of this most diverging sequence; and \code{MostDivSeq}, the most divergent sequence of characters, based on DeltaMatrix and MatrixMaxDivergence values.
}
\description{
Calculate the most divergent score and sequence for one sequence of characters
}
\examples{
SEQ1= c("A", "C","G","T","A","T","G","C","A","T")
MostDivergentScore(SEQ=SEQ1,MatrixMaxDivergence=MatrixMaxDivergence,DeltaMatrix = DeltaMatrix)
# $Length
# [1] 10
#
# $max.delta.vec
# A    C    G    T    A    T    G    C    A    T
# 0.11 0.14 0.13 0.12 0.11 0.12 0.13 0.14 0.11 0.12
#
# $Score
# [1] 22.77254
#
# $MostDivSeq
# [1] "-" "-" "-" "-" "-" "-" "-" "-" "-" "-"
}
