% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\title{Mutate the bases in a DNA sequence (only SNVs)}
\usage{
mutate(S, Nmut = 3)
}
\arguments{
\item{S}{Sequence of bases, as a vector of bases.}

\item{Nmut}{number of mutations (SNVs) in the sequence}
}
\value{
A vector of length(S) containing Nmut mutated bases
}
\description{
Mutate the bases in a DNA sequence (only SNVs)
}
\examples{
Seqi <-  paste(rep(c("A","T","G","C"),25))
Seqm <- mutate(Seqi,Nmut=10)
table(Seqi==Seqm)
}
