% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumSqrtScore.R
\name{sumSqrtScore}
\alias{sumSqrtScore}
\title{Sum of square roots of delta values}
\usage{
sumSqrtScore(DeltaVector = c(0.1, rep(0, 10)))
}
\arguments{
\item{DeltaVector}{The vector of delta values as compared to the aligned reference sequence.}
}
\value{
The score (numeric value) of the vector, as the sum of square root of each delta value. This is the unstandardized calculation. The score is not bound
}
\description{
Sum of square roots of delta values
}
\examples{

sumSqrtScore(rep(0,10000)) # 666716.5  score of a reference (deltas are all 0) that is 10000-nt long
sumSqrtScore(rep(0,100)) #  671.4629  score of a reference that is 100-nt long
sumSqrtScore(rep(0,10)) #  22.46828  score of a reference that is 10-nt long
}
